pytest(fix[fixtures]): move control fixtures into plugin
why: keep control-mode test helpers available when libtmux is used as a plugin dependency.
what:
- Relocate control_sandbox/control_client_logs fixtures into pytest_plugin
- Drop duplicate fixtures from top-level conftest
